Reorganize integer functions.
control.py was renamed procedural.py
