+++ mathics/builtin/procedural.py
@@ -3,7 +3,7 @@
 """
 Procedural Programming
 
-Procedural programming is a programming paradigm, derived from imperative programming, based on the concept of the procedure call.
+Procedural programming is a programming paradigm, derived from imperative programming, based on the concept of the procedure call. This term is sometimes compared and contrasted with Functional Programming.
 
 Procedures (a type of routine or subroutine) simply contain a series of computational steps to be carried out. Any given procedure might be called at any point during a program's execution, including by other procedures or itself.
