+++ setup.py
@@ -140,6 +140,7 @@ setup(
         "mathics.builtin.drawing",
         "mathics.builtin.fileformats",
         "mathics.builtin.files_io",
+        "mathics.builtin.intfns",
         "mathics.builtin.numbers",
         "mathics.builtin.numpy_utils",
         "mathics.builtin.pymimesniffer",
